Added Tilemap

+++ Project/Button.h
@@ -17,7 +17,7 @@ enum button_states { BTN_IDLE = 0, BTN_HOVER, BTN_ACTIVE };
 class Button
 {
 private:
-	short unsigned buttonStates;
+	unsigned buttonStates;
 	
 	sf::RectangleShape shape;
 	sf::Font* font;
